Move highlighter to new namespace
